setPremio

+++ database/migrations/2023_11_30_154240_create_premios_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
 
             $table->string('name');
             $table->text('detalle')->nullable();
-            $table->integer('puntos');
+            $table->decimal('puntos', 10, 2);
 
             $table->timestamps();
         });
